+++ docs/zapfdingbats.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 from itertools import chain
 from string import digits, ascii_uppercase, ascii_lowercase
-import fpdf
+from fpdf import FPDF, XPos, YPos
 
-pdf = fpdf.FPDF()
+pdf = FPDF()
 pdf.add_page()
 pdf.set_font("helvetica", size=30)
 pdf.cell(w=pdf.epw, txt="ZapfDingbats", align="C")
@@ -13,7 +13,7 @@ for i, char in enumerate(
     + digits
     + ":;<=>?@"
     + ascii_uppercase
-    + "[\]^_`"
+    + "[\\]^_`"
     + ascii_lowercase
     + "{|}~"
 ):
@@ -23,7 +23,7 @@ for i, char in enumerate(
     pdf.set_font("helvetica", size=30)
     pdf.cell(txt=char + " = ")
     pdf.set_font("zapfdingbats", size=30)
-    pdf.cell(txt=char, ln=1)
+    pdf.cell(txt=char, new_x=XPos.LMARGIN, new_y=YPos.NEXT)
 for i, n in enumerate(chain(range(0x80, 0x8E), range(0xA1, 0xF0), range(0xF1, 0xFF))):
     if i % 26 == 0:
         col = i // 26
@@ -34,5 +34,5 @@ for i, n in enumerate(chain(range(0x80, 0x8E), range(0xA1, 0xF0), range(0xF1, 0x
     pdf.set_font("helvetica", size=30)
     pdf.cell(txt=f"\\u00{hex(n)[2:]} = ")
     pdf.set_font("zapfdingbats", size=30)
-    pdf.cell(txt=chr(n), ln=1)
+    pdf.cell(txt=chr(n), new_x=XPos.LMARGIN, new_y=YPos.NEXT)
 pdf.output("zapfdingbats.pdf")